refactor out deliveries
